requirements.txt
Django>=3.2.4,<3.3.0
djangorestframework>=3.12.4,<3.13.0
psycopg2>=2.9.1,<3.0.0
flake8>=3.9.2,<4.0.0
django-rest-swagger>=2.2.0,<2.3.0